refactor: simplifies layout service
This commit reworks the logic in
layout service in order to make it
easier to understand.  The inital
logic was pulled from
angular-json-schema-form, but
was proving to be difficult to follow
as more functioanlity was added.
This will hopefully break up the code
a bit and make it more readable and
maintainable