feat: button component (#10)
* feat: adds simple button

Allows user to specify layotu to
show a simple button of type
button, reset, submit or image.

* feat: button with template

Allows specification of HTML
content to be used when rendering a button.  Type can be one of
button, reset or submit

* tests: adds missing test cases

* docs: adds missing documentation

* tests: add missing test case

* docs: Added missing docs

Also fixed some lint issues

* feat(button): Adds click handling

Adds ability to respond to custom
click handling of buttons

* tests: added missing test case

* refactor(button): cleaner abstraction of events

Removes the old event handing via the jsf service
in favor of using an EventPlugin.  This allows devs
to wire up any event to the jsf component and
listen for that event from a target with associated
layoutItem.

* cleanup: exports and unused code

exports the DataEvent type globally
and removes obsolete test code